requirements.txt
yfinance== 0.2.28 
streamlit ==1.26.0 
pandas ==2.1.0 
matplotlib ==3.7.2 
mysql-connector ==2.2.9 
plotly ==5.16.1 
deta
python-dotenv
quantstats
imgkit